@@ -3,12 +3,15 @@ import { toNumber } from '@sensenet/client-utils'
 import { File as SnFile } from '@sensenet/default-content-types'
 import {
   Annotation,
+  Comment,
+  CommentWithoutCreatedByAndId,
+  DocumentData,
   DocumentViewerSettings,
   Highlight,
   PreviewImageData,
   Redaction,
   Shape,
-} from '@sensenet/document-viewer-react/dist/models'
+} from '@sensenet/document-viewer-react'
 import { v1 } from 'uuid'
 
 /**
@@ -127,4 +130,43 @@ export const getViewerSettings = (repo: Repository) =>
       }
       return allPreviews
     },
+    commentActions: {
+      addPreviewComment: async (documentData, comment) => {
+        const response = await repo.executeAction<CommentWithoutCreatedByAndId, Comment>({
+          idOrPath: documentData.idOrPath,
+          body: comment,
+          name: 'AddPreviewComment',
+          method: 'POST',
+        })
+        return [response].map(changeCreatedByUrlToCurrent(documentData))[0]
+      },
+      deletePreviewComment: async (documentData, id) => {
+        const response = await repo.executeAction<any, { modified: boolean }>({
+          idOrPath: documentData.idOrPath,
+          body: { id },
+          name: 'DeletePreviewComment',
+          method: 'POST',
+        })
+        return response
+      },
+      getPreviewComments: async (documentData, page) => {
+        const response = await repo.executeAction<any, Comment[]>({
+          idOrPath: documentData.idOrPath,
+          name: 'GetPreviewComments',
+          method: 'GET',
+          oDataOptions: {
+            page,
+          } as any,
+        })
+        return response.map(changeCreatedByUrlToCurrent(documentData))
+      },
+    },
   })
+function changeCreatedByUrlToCurrent(documentData: DocumentData): (value: Comment) => Comment {
+  return comment => {
+    return {
+      ...comment,
+      createdBy: { ...comment.createdBy, avatarUrl: `${documentData.hostName}${comment.createdBy.avatarUrl}` },
+    }
+  }
+}
