@@ -1,18 +1,17 @@
 import { Injector } from '@furystack/inject'
-import { sensenetDocumentViewerReducer } from '@sensenet/document-viewer-react/dist/store'
+import { commentsStateReducer, sensenetDocumentViewerReducer } from '@sensenet/document-viewer-react'
 import { applyMiddleware, combineReducers, compose, createStore } from 'redux'
 import { ReduxDiMiddleware } from 'redux-di-middleware'
 import { commandPalette } from './CommandPalette'
 import { editContent } from './EditContent'
-
-const sensenetDocumentViewer = sensenetDocumentViewerReducer
 
 const composeEnhancers = (window as any).__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
 
 export const rootReducer = combineReducers({
   commandPalette,
   editContent,
-  sensenetDocumentViewer,
+  sensenetDocumentViewer: sensenetDocumentViewerReducer,
+  comments: commentsStateReducer,
 })
 export type rootStateType = ReturnType<typeof rootReducer>
 export const diMiddleware = new ReduxDiMiddleware(new Injector())
