@@ -3,6 +3,7 @@ import MuiThemeProvider from '@material-ui/core/styles/MuiThemeProvider'
 import Typography from '@material-ui/core/Typography'
 
 import {
+  defaultTheme,
   DocumentTitlePager,
   DocumentViewer,
   Download,
@@ -11,10 +12,10 @@ import {
   RotateActivePages,
   SearchBar,
   Share,
+  ToggleCommentsWidget,
   ToggleThumbnailsWidget,
   ZoomInOutWidget,
-} from '@sensenet/document-viewer-react/dist/components'
-import { exampleTheme } from '@sensenet/document-viewer-react/dist/ExampleAppLayout'
+} from '@sensenet/document-viewer-react'
 import { Icon, iconType } from '@sensenet/icons-react'
 import { compile } from 'path-to-regexp'
 import * as React from 'react'
@@ -125,7 +126,7 @@ export class DmsViewerComponent extends React.Component<
           }}
         />
         <div style={{ position: 'fixed', top: 0, left: 0, width: '100%', height: '100%' }}>
-          <MuiThemeProvider theme={exampleTheme}>
+          <MuiThemeProvider theme={defaultTheme}>
             <DocumentViewer
               documentIdOrPath={this.state.documentId}
               hostName={this.props.hostName}
@@ -160,7 +161,8 @@ export class DmsViewerComponent extends React.Component<
                           <RotateActivePages />
                         </div>
                         <DocumentTitlePager />
-                        <div style={{ flexShrink: 0 }}>
+                        <div style={{ display: 'flex', flexShrink: 0 }}>
+                          <ToggleCommentsWidget />
                           <SearchBar />
                         </div>
                       </LayoutAppBar>
