@@ -1,15 +1,15 @@
-import { sensenetDocumentViewerReducer } from '@sensenet/document-viewer-react/dist/store'
+import { commentsStateReducer, sensenetDocumentViewerReducer } from '@sensenet/document-viewer-react'
 import { Reducers } from '@sensenet/redux'
 import { combineReducers } from 'redux'
 import { dms } from '../Reducers'
 
 const sensenet = Reducers.sensenet
-const sensenetDocumentViewer = sensenetDocumentViewerReducer
 
 export const rootReducer = combineReducers({
   sensenet,
   dms,
-  sensenetDocumentViewer,
+  sensenetDocumentViewer: sensenetDocumentViewerReducer,
+  comments: commentsStateReducer,
 })
 
 export type rootStateType = ReturnType<typeof rootReducer>
