@@ -1,14 +1,10 @@
+import { Repository } from '@sensenet/client-core'
+import { BrowseView, EditView, NewViewComponent } from '@sensenet/controls-react/src'
+import { File, Folder } from '@sensenet/default-content-types'
 import { withA11y } from '@storybook/addon-a11y'
 import { object, text, withKnobs } from '@storybook/addon-knobs'
 import { storiesOf } from '@storybook/react'
 import React from 'react'
-
-import { Repository } from '@sensenet/client-core'
-import { File, Folder } from '@sensenet/default-content-types'
-import { Reducers } from '@sensenet/redux'
-
-import { BrowseView, EditView, NewViewComponent } from '@sensenet/controls-react/src'
-import { combineReducers, createStore } from 'redux'
 
 export const customSchema = [
   {
@@ -1538,9 +1534,6 @@ export const testFile = {
 const newViewNotes = require('../notes/viewcontrols/NewView.md')
 const editViewNotes = require('../notes/viewcontrols/EditView.md')
 const browseViewNotes = require('../notes/viewcontrols/BrowseView.md')
-
-const sensenet = Reducers.sensenet
-export const testStore = createStore(combineReducers({ sensenet }))
 
 storiesOf('ViewControls', module)
   .addDecorator(withKnobs)
