@@ -1,10 +1,15 @@
 import Drawer from '@material-ui/core/Drawer'
 import { SlideProps } from '@material-ui/core/Slide'
+import Typography from '@material-ui/core/Typography'
 import React = require('react')
 import { connect } from 'react-redux'
+import { DraftCommentMarker } from '../models'
 import { componentType } from '../services'
-import { RootReducerType, setActivePages, setThumbnails } from '../store'
-import { PageList } from './'
+import { createComment, RootReducerType, setActivePages, setThumbnails } from '../store'
+import { getComments, setSelectedCommentId } from '../store/Comments'
+import { CommentsContainer, PageList } from './'
+import Comment from './comment/Comment'
+import { CreateComment } from './comment/CreateComment'
 
 /**
  * maps state fields from the store to component props
@@ -16,7 +21,10 @@ const mapStateToProps = (state: RootReducerType) => {
     zoomMode: state.sensenetDocumentViewer.viewer.zoomMode,
     customZoomLevel: state.sensenetDocumentViewer.viewer.customZoomLevel,
     showThumbnails: state.sensenetDocumentViewer.viewer.showThumbnails,
+    showComments: state.sensenetDocumentViewer.viewer.showComments,
+    comments: state.comments.items,
     fitRelativeZoomLevel: state.sensenetDocumentViewer.viewer.fitRelativeZoomLevel,
+    localization: state.sensenetDocumentViewer.localization,
   }
 }
 
@@ -26,7 +34,10 @@ const mapStateToProps = (state: RootReducerType) => {
  */
 const mapDispatchToProps = {
   setActivePages,
+  getComments,
   setThumbnails,
+  createComment,
+  setSelectedCommentId,
 }
 
 /** Props definition for the Document Viewer layout */
@@ -36,6 +47,8 @@ export interface DocumentLayoutOwnProps {
 
 /** State type definition for the DocumentViewerLayout component */
 export interface DocumentLayoutState {
+  isPlacingCommentMarker: boolean
+  draftCommentMarker?: DraftCommentMarker
   activePage?: number
   thumbnaislVisibility: boolean
 }
@@ -47,8 +60,25 @@ export class DocumentViewerLayoutComponent extends React.Component<
   componentType<typeof mapStateToProps, typeof mapDispatchToProps, DocumentLayoutOwnProps>,
   DocumentLayoutState
 > {
-  /** the component state */
-  public state = { activePage: 1, thumbnaislVisibility: this.props.showThumbnails }
+  constructor(props: DocumentViewerLayoutComponent['props']) {
+    super(props)
+    this.state = {
+      activePage: 1,
+      thumbnaislVisibility: this.props.showThumbnails,
+      isPlacingCommentMarker: false,
+    }
+    this.createComment = this.createComment.bind(this)
+    this.handleKeyUp = this.handleKeyUp.bind(this)
+    this.handleMarkerCreation = this.handleMarkerCreation.bind(this)
+  }
+
+  public componentDidMount() {
+    document.addEventListener('keyup', this.handleKeyUp)
+  }
+
+  public componentWillUnmount() {
+    document.removeEventListener('keyup', this.handleKeyUp)
+  }
 
   /** scrolls the viewer to focus to the page with the provided index */
   public scrollTo(index: number, smoothScroll: boolean = true) {
@@ -75,8 +105,35 @@ export class DocumentViewerLayoutComponent extends React.Component<
 
       if (this.props.activePages[0] !== index) {
         this.props.setActivePages([index])
+        this.props.getComments()
       }
     })
+  }
+
+  private handleMarkerCreation(draftCommentMarker: DraftCommentMarker) {
+    this.setState({ ...this.state, draftCommentMarker })
+  }
+
+  private createComment(text: string) {
+    if (!this.state.draftCommentMarker || !this.state.activePage) {
+      return
+    }
+    this.props.createComment({ page: this.state.activePage, text, ...this.state.draftCommentMarker })
+    this.toggleIsPlacingCommentMarker(false)
+  }
+
+  private toggleIsPlacingCommentMarker(isPlacingCommentMarker = !this.state.isPlacingCommentMarker) {
+    this.setState({ ...this.state, isPlacingCommentMarker, draftCommentMarker: undefined })
+  }
+
+  private handleKeyUp(ev: KeyboardEvent) {
+    if (ev.key !== 'Escape') {
+      return
+    }
+    if (this.state.isPlacingCommentMarker) {
+      this.toggleIsPlacingCommentMarker(false)
+    }
+    this.props.setSelectedCommentId('')
   }
 
   /** triggered when the component will receive props */
@@ -86,6 +143,7 @@ export class DocumentViewerLayoutComponent extends React.Component<
       this.props.fitRelativeZoomLevel !== newProps.fitRelativeZoomLevel
     ) {
       this.scrollTo(newProps.activePages[0], this.props.fitRelativeZoomLevel === newProps.fitRelativeZoomLevel)
+      this.toggleIsPlacingCommentMarker(false)
     }
     if (this.props.showThumbnails !== newProps.showThumbnails) {
       if (newProps.showThumbnails) {
@@ -157,6 +215,8 @@ export class DocumentViewerLayoutComponent extends React.Component<
             />
           </Drawer>
           <PageList
+            handleMarkerCreation={this.handleMarkerCreation}
+            isPlacingCommentMarker={this.state.isPlacingCommentMarker}
             showWidgets={true}
             id="sn-document-viewer-pages"
             zoomMode={this.props.zoomMode}
@@ -169,6 +229,33 @@ export class DocumentViewerLayoutComponent extends React.Component<
             padding={8}
             activePage={this.state.activePage}
           />
+          <Drawer
+            variant={'persistent'}
+            open={this.props.showComments}
+            anchor="right"
+            SlideProps={this.props.drawerSlideProps}
+            PaperProps={{
+              style: {
+                position: 'relative',
+                width: this.props.showComments ? '340px' : 0,
+                height: '100%',
+                overflow: 'hidden',
+              },
+            }}>
+            <CommentsContainer>
+              <Typography variant="h4">{this.props.localization.commentSideBarTitle}</Typography>
+              <CreateComment
+                draftCommentMarker={this.state.draftCommentMarker}
+                handlePlaceMarkerClick={() => this.toggleIsPlacingCommentMarker()}
+                isPlacingMarker={this.state.isPlacingCommentMarker}
+                localization={this.props.localization}
+                createComment={this.createComment}
+              />
+              {this.props.comments.map(comment => (
+                <Comment key={comment.id} {...comment} />
+              ))}
+            </CommentsContainer>
+          </Drawer>
         </div>
       </div>
     )
