@@ -1,9 +1,11 @@
-import * as React from 'react'
+import React from 'react'
 import { connect } from 'react-redux'
-import { Annotation, Highlight, PreviewImageData, Redaction, Shape, Shapes } from '../../models'
+import { Annotation, DraftCommentMarker, Highlight, PreviewImageData, Redaction, Shape, Shapes } from '../../models'
 import { componentType, Dimensions } from '../../services'
 import { RootReducerType, updateShapeData } from '../../store'
+import { setSelectedCommentId } from '../../store/Comments'
 import { ShapeAnnotation, ShapeHighlight, ShapeRedaction } from './Shape'
+import { CommentMarker, ShapesContainer } from './style'
 
 /**
  * Defined the component's own properties
@@ -12,6 +14,7 @@ export interface OwnProps {
   viewPort: Dimensions
   page: PreviewImageData
   zoomRatio: number
+  draftCommentMarker?: DraftCommentMarker
 }
 
 /**
@@ -19,9 +22,16 @@ export interface OwnProps {
  * @param state the redux state
  */
 export const mapStateToProps = (state: RootReducerType, ownProps: OwnProps) => {
+  const commentMarkers = state.comments.items
+    .filter(comment => comment.page === ownProps.page.Index)
+    .map(comment => {
+      return { x: comment.x, y: comment.y, id: comment.id }
+    })
+  ownProps.draftCommentMarker && commentMarkers.push(ownProps.draftCommentMarker)
   return {
     showShapes: state.sensenetDocumentViewer.viewer.showShapes,
     showRedactions: state.sensenetDocumentViewer.viewer.showRedaction,
+    showComments: state.sensenetDocumentViewer.viewer.showComments,
     canHideRedactions: state.sensenetDocumentViewer.documentState.canHideRedaction,
     redactions: state.sensenetDocumentViewer.documentState.document.shapes.redactions.filter(
       r => r.imageIndex === ownProps.page.Index,
@@ -32,6 +42,8 @@ export const mapStateToProps = (state: RootReducerType, ownProps: OwnProps) => {
     annotations: state.sensenetDocumentViewer.documentState.document.shapes.annotations.filter(
       r => r.imageIndex === ownProps.page.Index,
     ) as Annotation[],
+    commentMarkers,
+    selectedCommentId: state.comments.selectedCommentId,
     canEdit: state.sensenetDocumentViewer.documentState.canEdit,
   }
 }
@@ -42,6 +54,7 @@ export const mapStateToProps = (state: RootReducerType, ownProps: OwnProps) => {
  */
 export const mapDispatchToProps = {
   updateShapeData,
+  setSelectedCommentId,
 }
 
 /**
@@ -73,50 +86,59 @@ export class ShapesComponent extends React.Component<
    */
   public render() {
     return (
-      <div
-        onDrop={ev => this.onDrop(ev, this.props.page)}
-        onDragOver={ev => ev.preventDefault()}
-        style={{ position: 'absolute', top: 0, left: 0, width: '100%', height: '100%', zIndex: 1 }}>
-        {this.props.canHideRedactions &&
-          this.props.showRedactions &&
-          this.props.redactions.map((redaction, index) => {
-            return (
-              <ShapeRedaction
-                shapeType="redactions"
-                shape={redaction}
-                canEdit={this.props.canEdit}
-                zoomRatio={this.props.zoomRatio}
-                key={index}
-              />
-            )
-          })}
+      <ShapesContainer>
+        {this.props.showComments &&
+          this.props.commentMarkers.map(marker => (
+            <CommentMarker
+              onClick={() => this.props.setSelectedCommentId(marker.id)}
+              isSelected={marker.id === this.props.selectedCommentId}
+              zoomRatio={this.props.zoomRatio}
+              marker={marker}
+              key={marker.id}
+            />
+          ))}
+        <div onDrop={ev => this.onDrop(ev, this.props.page)} onDragOver={ev => ev.preventDefault()}>
+          {this.props.canHideRedactions &&
+            this.props.showRedactions &&
+            this.props.redactions.map((redaction, index) => {
+              return (
+                <ShapeRedaction
+                  shapeType="redactions"
+                  shape={redaction}
+                  canEdit={this.props.canEdit}
+                  zoomRatio={this.props.zoomRatio}
+                  key={index}
+                />
+              )
+            })}
 
-        {this.props.showShapes &&
-          this.props.annotations.map((annotation, index) => {
-            return (
-              <ShapeAnnotation
-                shapeType="annotations"
-                shape={annotation}
-                canEdit={this.props.canEdit}
-                zoomRatio={this.props.zoomRatio}
-                key={index}
-              />
-            )
-          })}
+          {this.props.showShapes &&
+            this.props.annotations.map((annotation, index) => {
+              return (
+                <ShapeAnnotation
+                  shapeType="annotations"
+                  shape={annotation}
+                  canEdit={this.props.canEdit}
+                  zoomRatio={this.props.zoomRatio}
+                  key={index}
+                />
+              )
+            })}
 
-        {this.props.showShapes &&
-          this.props.highlights.map((highlight, index) => {
-            return (
-              <ShapeHighlight
-                shapeType="highlights"
-                shape={highlight}
-                canEdit={this.props.canEdit}
-                zoomRatio={this.props.zoomRatio}
-                key={index}
-              />
-            )
-          })}
-      </div>
+          {this.props.showShapes &&
+            this.props.highlights.map((highlight, index) => {
+              return (
+                <ShapeHighlight
+                  shapeType="highlights"
+                  shape={highlight}
+                  canEdit={this.props.canEdit}
+                  zoomRatio={this.props.zoomRatio}
+                  key={index}
+                />
+              )
+            })}
+        </div>
+      </ShapesContainer>
     )
   }
 }
