@@ -2,6 +2,7 @@ import { combineReducers } from 'redux'
 import { documentStateReducer, DocumentStateType } from './Document'
 import { localizationReducer, LocalizationStateType } from './Localization'
 
+import { CommentsState, commentsStateReducer } from './Comments'
 import { previewImagesReducer, PreviewImagesStateType } from './PreviewImages'
 import { viewerStateReducer, ViewerStateType } from './Viewer'
 
@@ -9,6 +10,7 @@ import { viewerStateReducer, ViewerStateType } from './Viewer'
  * Type definitions for the sensenet document viewer root reducer
  */
 export interface RootReducerType {
+  comments: CommentsState
   sensenetDocumentViewer: {
     /**
      * Reducer for the document state
@@ -33,18 +35,6 @@ export interface RootReducerType {
 }
 
 /**
- * The root reducer instance with the sensenetDocumentViewer reducer instance
- */
-export const rootReducer = combineReducers<RootReducerType>({
-  sensenetDocumentViewer: combineReducers({
-    documentState: documentStateReducer,
-    previewImages: previewImagesReducer,
-    viewer: viewerStateReducer,
-    localization: localizationReducer,
-  }),
-})
-
-/**
  * The sensenetDocumentViewerReducer instance
  */
 export const sensenetDocumentViewerReducer = combineReducers({
@@ -52,4 +42,12 @@ export const sensenetDocumentViewerReducer = combineReducers({
   previewImages: previewImagesReducer,
   viewer: viewerStateReducer,
   localization: localizationReducer,
+})
+
+/**
+ * The root reducer instance with the sensenetDocumentViewer reducer instance
+ */
+export const rootReducer = combineReducers<RootReducerType>({
+  comments: commentsStateReducer,
+  sensenetDocumentViewer: sensenetDocumentViewerReducer,
 })
